refactor mixin

+++ Nodes/bone_transform_nodes.py
@@ -233,7 +233,7 @@ class DefineBoneTransform(_BoneTransform):
         # Output End (UI)
         out_end = self.outputs.get("End")
         if out_end:
-            try: out_end.default_value = float(start + max(0, duration))
+            try: out_end.default_value = int(start + max(0, duration))
             except Exception: pass
 
         # If before start: clear cache entry and do nothing
